feat: 🎸 Added Cancel Confirmation Dialog modal
 In an edit scenario we pull up a dialog with all of the initial values
set. If we change a value we can now submit the values with the one
change. We can also Cancel out of the dialog at any point. If we have
dirtied the form by changing a value and then click cancel/close our
form should check that we didn't want to Submit/Save that value. We
should get some kind of "Are you sure you want to cancel without saving
your changes?

BREAKING CHANGE: 🧨 No

✅ Closes: #60