refactor: 💡 remove unnecessary prop
Recently we added a prop to hide the header, but actually we do not need
this prop and we can simply use the title to determine if a header needs
to be displayed. If it has a title property set, we show the header, if
there's not a title set than we simply skip rendering it.