fix: 🐛 SQFormAutocomplete fix JSS warning
While running tests for SQFormAutocomplete, the same warning is thrown
every test. Warning: [JSS] Cannot set property 'parentStyleSheet' of
undefined. This was solved by returning undefined instead of empty
string so that the conditional styles would not be included if they were
not needed (instead of setting them to empty string)

✅ Closes: #510